+++ Assets/[0]Game/[0]Code/Battle/Enemies/Dragon.cs
@@ -12,7 +12,7 @@ namespace Game
         {
             if (eventName == "Damage")
             {
-                yield return _damageEvent.AwaitEvent(this, value);
+                yield return _damageEvent.AwaitEvent(this, (int)value);
             }
         }
     }
