+++ Assets/[0]Game/[0]Code/Battle/Enemies/FakeHero.cs
@@ -23,7 +23,7 @@ namespace Game
 
             if (eventName == "Damage")
             {
-                yield return _damageEvent.AwaitEvent(this, value);
+                yield return _damageEvent.AwaitEvent(this, (int)value);
             }
 
             if (eventName == "EndBattle")
