+++ Assets/[0]Game/[0]Code/Battle/Enemies/Slime.cs
@@ -62,7 +62,7 @@ namespace Game
 
             if (eventName == "Damage")
             {
-                yield return _damageEvent.AwaitEvent(this, value);
+                yield return _damageEvent.AwaitEvent(this, (int)value);
             }
             
             if (eventName == "EndBattle")
